requirements.txt
numpy
torch
torchvision
rasterio
Pillow
matplotlib
tqdm
PyYAML
pathlib
h5py
scikit-learn